requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9

sqlalchemy==2.0.36

apscheduler==3.10.4
pytz==2024.2

anthropic==0.40.0
python-dotenv==1.0.1
pydantic-settings==2.6.1

tweepy==4.14.0
newsapi-python==0.2.7
feedparser==6.0.11
httpx==0.27.2
requests==2.32.3
